Rename "InertiaLink" file to "Link"
